\name{cmdpipe}
\alias{cmdpipe}
\docType{data}
\title{
data name/kind
}
\description{
A concise (1-5 lines) description of the dataset.
}
\usage{data("cmdpipe")}
\format{
  The format is:
 'pipe' int 3
 - attr(*, "conn_id")=<externalptr> 
}
\details{
If necessary, more details than the __description__ above
}
\source{
reference to a publication or URL from which the data were obtained
}
\references{
possibly secondary sources and usages
}
\examples{
data(cmdpipe)
## maybe str(cmdpipe) ; plot(cmdpipe) ...
}
\keyword{datasets}
