\name{moduleInitR-package}
\alias{moduleInitR-package}
\alias{moduleInitR}
\docType{package}
\title{
\packageTitle{moduleInitR}
}
\description{
\packageDescription{moduleInitR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{moduleInitR}
\packageIndices{moduleInitR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{moduleInitR}

Maintainer: \packageMaintainer{moduleInitR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
